feat: bug fix lexer for parsing large integers

+++ crates/cyrusc_cir/src/lib.rs
@@ -195,7 +195,7 @@ pub struct CIRLiteral {
 
 #[derive(Debug, Clone)]
 pub enum CIRLiteralKind {
-    Integer(i64, bool),
+    Integer(i128, bool),
     Float(f64),
     Bool(bool),
     Char(char),
